pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rakkib"
version = "2.0.0a0"
description = "Rakkib v2 — wizard-driven server bootstrap"
requires-python = ">=3.9"
dependencies = [
  "click>=8.0",
  "fastapi>=0.115",
  "pyyaml>=6.0",
  "rich>=13.0",
  "jinja2>=3.1",
  "questionary>=2.0",
  # qrcode 8.x removed the 'cli' extra; we only need the library.
  "qrcode>=7.4",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
test = [
  "httpx>=0.28",
  "pytest>=8.0",
]

[project.scripts]
rakkib = "rakkib.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rakkib = ["data/**/*"]
